ARM: dts: qcom: msm8974: initial device tree for Samsung Galaxy S5 China
Add initial device tree for Samsung Galaxy S5 China. This device is
almost the same as Samsung Galaxy S5 (klte). The only difference is the
gpio pins of i2c-gpio-led and qcom-msm8974-samsung-klte.dts was used as
an example. There are still other differences but affected devices all
lack of mainline drivers and are not in the klte DT.

Signed-off-by: Rongrong Zhang <i@rong.moe>
Signed-off-by: xtex <xtexchooser@duck.com>